Summary of Changes:
When the user uploads an image, the image is now stored in the database instead of being stored on the file system and linked. Note that this includes the default image which is still used if no twitter Url is provided and no image is uploaded.

A link is still used for users which specify a twitter URL.
An 'Image' property was added to all objects that already contain an ImageURL property. The code decides what to use/show depending on the value of these properties. Generally the Image property takes precedence if available.

Existing users whose image was previously stored on the file system and is now lost will need to re-upload their image in order for it to be stored/used correctly by the new site.

In conjunction with this change some parts of the code that previously would not show anything in the case that no image is available will now show the default image, other than this there is no visible UI changes for this fix.

Related Bug Fix: The external login page now handles uploaded avatar images correctly, previously an uploaded image file would be ignored & the default or twitter image (if specified) would be used.